requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
supabase==2.4.6
httpx==0.27.0
pydantic-settings==2.2.1